+++ HDSLedger/Service/src/main/java/pt/ulisboa/tecnico/hdsledger/service/models/Transaction.java
@@ -1,18 +1,16 @@
 package pt.ulisboa.tecnico.hdsledger.service.models;
 
 import com.google.gson.Gson;
-
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
+import pt.ulisboa.tecnico.hdsledger.utilities.Colors;
 
 public class Transaction {
     private final String sender;
     private final String receiver;
-    private final int amount;
+    private final double amount;
     private final int nonce;
     private final byte[] signature;
 
-    public Transaction(String sender, String receiver, Integer amount, Integer nonce, byte[] signature){
+    public Transaction(String sender, String receiver, double amount, Integer nonce, byte[] signature){
         this.sender = sender;
         this.receiver = receiver;
         this.amount = amount;
@@ -20,7 +18,7 @@ public class Transaction {
         this.signature = signature;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
